allow floats for reg

+++ app/configuration.py
@@ -10,6 +10,6 @@ def parse_configuration(configuration):
 
     ar_config = configuration.get("autoRegression", {})
     ar_lags = int(ar_config.get("lags", 0))
-    ar_regularization = int(ar_config.get("regularization", 0))
+    ar_regularization = float(ar_config.get("regularization", 0))
 
     return epochs, forecasts, ar_lags, ar_regularization
